+++ modules/data_integrity_flag/templates/menu_data_integrity_flag.tpl
@@ -1,4 +1,4 @@
 {*
-  Everything is rendered in index.js now!
+  Everything is rendered in DataIntegrityFlag.js now!
   I'm here because currently Loris breaks without me :(
 *}
